Multiple database modifications and warning icon in stock component
The group short name was removed from the groups table because it wasn't used anymore.

A settings table was created. This table contain all settings that belong to a group and that can be updated by the administrator (for example, minimal stock alert number).

Consequently, a seeding for this table has been set up (the table only contains the stock_empty_alert setting), a model has been created and also migration.

And last, but not least, a warning icon will now be displayed when the stock for a specific product reaches the stock_empty_alert value.

Fixed #35